+++ lib/service/Groups.dart
@@ -33,6 +33,7 @@ import 'package:illinois/utils/Utils.dart';
 class Groups /* with Service */ {
 
   static const String notifyUserMembershipUpdated   = "edu.illinois.rokwire.groups.membership.updated";
+  static const String notifyGroupEventsUpdated   = "edu.illinois.rokwire.groups.events.updated";
   static const String notifyGroupCreated            = "edu.illinois.rokwire.group.created";
   static const String notifyGroupUpdated            = "edu.illinois.rokwire.group.updated";
   static const String notifyGroupDeleted            = "edu.illinois.rokwire.group.deleted";
@@ -389,18 +390,22 @@ class Groups /* with Service */ {
     return false;
   }
 
+  Future<void> updateGroupEvents(Event event){
+    return ExploreService().updateEvent(event).then((eventId) {
+      NotificationService().notify(Groups.notifyGroupEventsUpdated);
+    });
+  }
+
   Future<bool> deleteEventFromGroup({String groupId, String eventId}) async {
-    //TBD
-    return false;
+    await removeEventFromGroup(groupId: groupId, eventId: eventId);
+    bool deleteResult = await ExploreService().deleteEvent(eventId);
+    NotificationService().notify(Groups.notifyGroupEventsUpdated);
+    return deleteResult;
   }
 
   Future<Event> createGroupEvent(String groupId, Event event) async {
     //TBD
     return Future<Event>.delayed(Duration(seconds: 1), (){ return event; });
-  }
-
-  Future<bool> updateGroupEvents(String groupId, List<Event> events) async {
-    return Future<bool>.delayed(Duration(seconds: 1), (){ return true; });
   }
 
   // Event Comments